refactor: add api request for contract template
